+++ modules/globebrowsing/globes/chunkedlodglobe.cpp
@@ -251,10 +251,10 @@ float ChunkedLodGlobe::getHeight(glm::dvec3 position) const {
             continue;
         }
 
-        float sample0 = sample00 * (1.0 - samplePosFract.x) + sample10 * samplePosFract.x;
-        float sample1 = sample01 * (1.0 - samplePosFract.x) + sample11 * samplePosFract.x;
+        float sample0 = sample00 * (1.f - samplePosFract.x) + sample10 * samplePosFract.x;
+        float sample1 = sample01 * (1.f - samplePosFract.x) + sample11 * samplePosFract.x;
 
-        float sample = sample0 * (1.0 - samplePosFract.y) + sample1 * samplePosFract.y;
+        float sample = sample0 * (1.f - samplePosFract.y) + sample1 * samplePosFract.y;
 
         // Same as is used in the shader. This is not a perfect solution but
         // if the sample is actually a no-data-value (min_float) the interpolated
@@ -356,7 +356,9 @@ void ChunkedLodGlobe::debugRenderChunk(const Chunk& chunk, const glm::dmat4& mvp
 }
 
 void ChunkedLodGlobe::update(const UpdateData& data) {
-    setBoundingSphere(_owner.ellipsoid().maximumRadius() * data.modelTransform.scale);
+    setBoundingSphere(static_cast<float>(
+        _owner.ellipsoid().maximumRadius() * data.modelTransform.scale
+    ));
     _renderer->update();
 }
