+++ modules/globebrowsing/rendering/gpu/gpulayermanager.cpp
@@ -51,7 +51,12 @@ void GPULayerManager::bind(ghoul::opengl::ProgramObject* programObject,
     
     for (size_t i = 0; i < layerGroups.size(); ++i) {
         const std::string& nameBase = layergroupid::LAYER_GROUP_NAMES[i];
-        _gpuLayerGroups[i]->bind(programObject, *layerGroups[i], nameBase, i);
+        _gpuLayerGroups[i]->bind(
+            programObject,
+            *layerGroups[i],
+            nameBase,
+            static_cast<int>(i)
+        );
     }
 }
